requirements.txt
openai>=1.50.0
qdrant-client>=1.9.0
pypdf>=4.0.0
tiktoken>=0.7.0
python-dotenv>=1.0.1
pydantic-ai>=0.0.15
serpapi>=0.1.4
logfire>=4.15.1
